Clean up inputRule
